requirements.txt
langchain
langchain-community
langchain_groq
langchain_huggingface
streamlit
pandas
python-dotenv
sentence-transformers
faiss-cpu